+++ operator/api/v1beta1/aistore_webhook_test.go
@@ -1,0 +1,137 @@
+package v1beta1
+
+import (
+	"testing"
+
+	aisapc "github.com/NVIDIA/aistore/api/apc"
+	. "github.com/onsi/gomega"
+	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
+)
+
+func TestAIStoreValidateSize(t *testing.T) {
+	tests := []struct {
+		name       string // description of this test case
+		want       admission.Warnings
+		wantErr    bool
+		proxySize  *int32
+		targetSize *int32
+		size       *int32
+	}{
+		{
+			"Proxy size is -1 thus proxy autoscaling is true",
+			nil,
+			false,
+			aisapc.Ptr[int32](-1),
+			aisapc.Ptr[int32](1),
+			nil,
+		},
+		{
+			"target size is -1 thus target autoscaling is true",
+			nil,
+			false,
+			aisapc.Ptr[int32](1),
+			aisapc.Ptr[int32](-1),
+			nil,
+		},
+		{
+			" size is -1 thus autoscaling is true",
+			nil,
+			false,
+			nil,
+			nil,
+			aisapc.Ptr[int32](-1),
+		},
+		{
+			"autoscaling",
+			nil,
+			false,
+			aisapc.Ptr[int32](1),
+			aisapc.Ptr[int32](1),
+			aisapc.Ptr[int32](-1),
+		},
+		{
+			"not autoscaling",
+			nil,
+			false,
+			aisapc.Ptr[int32](1),
+			aisapc.Ptr[int32](1),
+			nil,
+		},
+		{
+			"not autoscaling with just size",
+			nil,
+			false,
+			nil,
+			nil,
+			aisapc.Ptr[int32](1),
+		},
+		{
+			"invalid size",
+			nil,
+			true,
+			nil,
+			nil,
+			aisapc.Ptr[int32](-2),
+		},
+		{
+			"invalid target size",
+			nil,
+			true,
+			aisapc.Ptr[int32](1),
+			aisapc.Ptr[int32](-2),
+			nil,
+		},
+		{
+			"invalid proxy size",
+			nil,
+			true,
+			aisapc.Ptr[int32](-2),
+			aisapc.Ptr[int32](1),
+			nil,
+		},
+		{
+			"invalid proxy size;0",
+			nil,
+			true,
+			aisapc.Ptr[int32](0),
+			aisapc.Ptr[int32](1),
+			nil,
+		},
+		{
+			"invalid target size;0",
+			nil,
+			true,
+			aisapc.Ptr[int32](1),
+			aisapc.Ptr[int32](0),
+			nil,
+		},
+		{
+			"invalid target size",
+			nil,
+			true,
+			aisapc.Ptr[int32](0),
+			aisapc.Ptr[int32](0),
+			aisapc.Ptr[int32](0),
+		},
+	}
+	for _, tt := range tests {
+		RegisterTestingT(t)
+		t.Run(tt.name, func(t *testing.T) {
+			var ais AIStore
+			ais.Spec.ProxySpec.Size = tt.proxySize
+			ais.Spec.TargetSpec.Size = tt.targetSize
+			ais.Spec.Size = tt.size
+			got, gotErr := ais.validateSize()
+			if gotErr != nil {
+				if !tt.wantErr {
+					t.Errorf("validateSize() failed: %v for test %s", gotErr, tt.name)
+				}
+				return
+			}
+			if tt.wantErr {
+				t.Fatal("validateSize() succeeded unexpectedly")
+			}
+			Expect(got).To(Equal(tt.want))
+		})
+	}
+}