+++ operator/pkg/resources/target/target_suite_test.go
@@ -1,0 +1,13 @@
+package target
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestCommon(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Target Suite")
+}